Update integration tests.

+++ src/integration-test/java/com/chumbok/multitenancy/TestApplication.java
@@ -3,16 +3,22 @@ package com.chumbok.multitenancy;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.PropertySource;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.Collections;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 @SpringBootApplication
 @PropertySource("classpath:application.yml")
@@ -32,8 +38,15 @@ public class TestApplication {
     }
 
     @GetMapping("/hello/{id}")
-    public void getHello(@PathVariable String id) {
-        helloRepository.findById(id);
+    public Hello getHello(@PathVariable String id) {
+
+        Optional<Hello> hello = helloRepository.findById(id);
+
+        if (!hello.isPresent()) {
+            throw new RuntimeException("Hello[" + id + "] does not exist.");
+        }
+
+        return hello.get();
     }
 
 
@@ -44,12 +57,34 @@ public class TestApplication {
 
     @PutMapping("/hello/{id}")
     public Hello updateHello(@PathVariable String id, @RequestBody Hello hello) {
-        return helloRepository.saveAndFlush(hello);
+
+        Optional<Hello> savedHello = helloRepository.findById(id);
+
+        if (!savedHello.isPresent()) {
+            throw new RuntimeException("Hello[" + id + "] does not exist.");
+        }
+
+        savedHello.get().setMessage(hello.getMessage());
+        return helloRepository.saveAndFlush(savedHello.get());
     }
 
     @DeleteMapping("/hello/{id}")
     public void deleteHello(@PathVariable String id) {
+
+        Optional<Hello> savedHello = helloRepository.findById(id);
+
+        if (!savedHello.isPresent()) {
+            throw new RuntimeException("Hello[" + id + "] does not exist.");
+        }
+
         helloRepository.deleteById(id);
+    }
+
+    // Returning INTERNAL_SERVER_ERROR since it just a test
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    @ExceptionHandler(RuntimeException.class)
+    public Map<String, String> handleException(RuntimeException ex) {
+        return Collections.singletonMap("message", ex.getMessage());
     }
 }
